requirements.txt
Django==3.1.4
psycopg2==2.8.6
django-autoslug==1.9.8
requests==2.25.0
dj-database-url==0.5.0
python-decouple==3.3
django-allauth==0.44.0
smarturls==0.1.9
# django-el-pagination==3.3.0
djangorestframework==3.12.2
pillow==8.0.1
# django-cors-headers==3.4.0: No instalado porque creo que no es necesario, por ahora, ya que es para intercambio de recursos